# docs/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM System - REST API Documentation",
        "description": "CRM System - REST API Documentation built by Victor Tavares",
        "contact": {
            "email": "victortavaresdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/users/create": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "operationId": "create-user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "example": "Amanda"
                                    },
                                    "last_name": {
                                        "example": "Torres"
                                    },
                                    "email": {
                                        "example": "amanda@gmail.com"
                                    },
                                    "password": {
                                        "example": "teste123"
                                    },
                                    "address": {
                                        "example": "Rua machado de assis"
                                    },
                                    "phone_number": {
                                        "example": "(11) 98080-4040"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user data",
                "operationId": "get-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}/update": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user data",
                "operationId": "update-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "example": "Amanda updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}/delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user data",
                "operationId": "delete-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Access token obtained from authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}